Version 1.4.1

+++ tests/epyccel/modules/generic_functions_2.py
@@ -171,7 +171,7 @@ def mix_complex_array_1(x, a):
 
 #$ header function dup_header(real)
 #$ header function dup_header(float64)
-@types('float64')
+@types('float')
 @types('float64')
 def dup_header(a):
     return a
